Move splitOverDays function into RangeUtils and refactor tests
